+++ internal/proxy/roundtripper.go
@@ -1,7 +1,6 @@
 package proxy
 
 import (
-	"context"
 	"crypto/tls"
 	"net"
 	"net/http"
@@ -27,30 +26,6 @@ type h2cTransportWrapper struct {
 
 	log  log.Logger
 	hook hook.Service
-}
-
-type Context struct {
-	ctx context.Context
-}
-
-func (c Context) Deadline() (time.Time, bool) {
-	return time.Time{}, false
-}
-
-func (c Context) Done() <-chan struct{} {
-	return nil
-}
-
-func (c Context) Err() error {
-	return nil
-}
-
-func (c Context) Value(key interface{}) interface{} {
-	return c.ctx.Value(key)
-}
-
-func WithoutCancel(ctx context.Context) context.Context {
-	return Context{ctx: ctx}
 }
 
 func (t *h2cTransportWrapper) RoundTrip(req *http.Request) (*http.Response, error) {
